PushMediaConstraints.java: Tweak image dimensions (maximum to 2160 pixels)
This PR tweaks the maximum image dimensions to fall in line with screen resolutions and video standards.

This is not a perfect solution, make image compression slightly less worse than without it.

Recently the maximum image width was reduced from 4096 pixels to 1600 pixels, the same number WhatsApp uses for years. 1600 isn't the most logical number. In this PR I used standard video formats as baseline, which most smartphone displays also use.

The maximum of 2160 is both the vertical resolution of a regular 4K-screen (3840x2160), but also the vertical resolution of a 18:9 1080p display (1080x2160). So it's high enough to share 1080p screenshots, even in the 18:9 ratio.

Another advantage that for pictures in the 3:2 aspect ratio 2160 is perfectly dividable: 2160/3*2 = 1440 pixels on the short side, while 1600 pixels tries to reduce it to 1066.67 pixels, mostly rounded to 1067.

The max resolution is a little higher and so will the file-size be, but that's were the fall-backs are for. These are also adjusted to common screen and video format resolutions.

------

Talking about fall-backs: Currently the encoder tries (as far as I understand it) to drop resolution when the maximum file size is exceeded. I don't think this would be the best way, I would suggest dropping the quality parameter in the encoder. By default JPEG quality 70 now is used, instead of dropping resolution dropping encoder quality to 60 or 50 should first be tried.