Add mobile call data usage option
This change adds an option to restrict calls from using mobile data. My wife
experienced a video call using 750 MB of data even though she was on WiFI. It
needs to be controllable in case you are on a metered plan since it might
otherwise make the call very expensive. My wife almost uninstalled Signal
because of this, but I got her to stay by promising I would fix it.

This change depends on my changes made to RingRTC. You need to make sure you
have the right version of RingRTC.