Show thread without messages
The current behavior is that threads without messages are removed/hidden
from the default and archived lists.
They can still be found in the NewConversationActivity, but the behavior
is confusing for users.
The user can already delete a conversation explicitly, so there's no need
to automatically delete the conversation when its last message is deleted.