Use Captcha-URL from BuildConfig
Actually the used URL is read from RegistrationConstants. But in the app's build.gradle we can define the URL for the recaptcha which is than generated in the builds config, but unused here. So it will be nice to really use the defined parameter from that config. Otherwise the definition in build.gradle is useless and can be removed there. I think using the BuildConfig would be the better solution.