Ensure last version code is stored in SharedPreferences.
`VersionTracker.updateLastSeenVersion` is the only place in the codebase
that sets the last version code in SharedPreferences. However, on the
initial run of the app, the preference won't be set. This results in the
`TextSecurePreferences.getLastVersionCode` using the current version
as the default value.

But this results in the `VersionTracker` code skipping the if-statement
block, so it doesn't set the last version code in the SharedPreferences.
What results is that we never have a value for the last version code, so
`TextSecurePreferences.getLastVersionCode` always defaults to the
current version code, and thus we always skip the if-statement in
`VersionTracker.updateLastSeenVersion`.

Fixes #11776