Adds content descriptions for in-call buttons
Fixes #9774

Previously, when in a WebRTC call and using Talkback, the dailer
buttons would read, from right to left:
- unlabelled button
- (on/off) switch
- (on/off) switch
- unlabelled button (when showing flip)

now, when in a WebRTC call and using Talkback, with video and audio
enabled, the dailer buttons will read (from right to left):
- end call button
- on unmute switch
- on camera switch
- flip button

when in a WebRTC call and using Talkback with video off and audio
muted, the dailer buttons will read (from right to left):
- end call button
- off unmute switch
- off camera switch
- flip button

I do think that it would likely be a better user experience if the
"flip" button read "flip to front camera" or "flip to back camera"
but it looks like that may be fairly complicated and above my
skills at this time. As it seems like this issue has been open for
a long time, maybe it's helpful to merge the initial PR, then have
someone else follow up later?