Adds content description for viewing user photo
fixes https://github.com/signalapp/Signal-Android/issues/12549

If a user has an image set, then the view showing their picture is
itself a button that allows you to view the image at a larger
size.

If a user does not have an image set, then the view is still a
button but the button won't do anything...

After this change, bringing Talkback focus to the user's profile
image will read "view contact photo" (even if there is no contact
photo to view). I did try to see if I could disambiguate whether
there was something to be viewed or not, but that proved a bit
beyond my time + skills right now, so hopefully it's more helpful
than harmful to have the view labelled this way.