Avoid dismissing `SimpleProgressDialog` a second time
In case `dismiss` (with delay) is called during an activity and the
activity is consequently destroyed, and calls `dismissNow` in the
process of its cleanup (i.e. in `onDestroy`), the delayed `dismiss`
callback could still try to dismiss the dialog, causing an
`IllegalArgumentException: View[…] not attached to window manager`.

To avoid this faulty behavior, the delayed callbacks are removed
whenever `dismissNow` is called.