Hide default 'Signal' text when calls in call log selected
Fixes issue #13260 where the user would see 'Signal' with "[n] selected" in the ActionBar. 

Achieved by first hiding supportActionBar when ActionMode starts (items selected) and showing supportActionBar when it ends (all items deselected). The supportActionBar also needs to be hidden again in onResume when the user has navigated back to the screen (but only if no items are selected, using viewModel.selectionStateSnapshot.isNotEmpty(...).