Use LinearLayout for attachment keyboard buttons
RecyclerView introduces excessive boilerplate for a container whose size
is roughly fixed at a small number and is generally expected to not need
recycling to be done. A simple horizontal LinearLayout wrapped in a
HorizontalScrollView for good measure simplifies the code and reduces
weird interactions with the recycler when the attachment keyboard is
resized as part of the smooth open/close animation that I am adding.