Fade out the fake keyboard when closing it
When we remove the fake keyboard fragment, use an animated transition to
make it look less choppy.

Also don't forcibly end the smooth transition for the fake keyboard
unless the animation is actually in progress. This prevents erroneously
setting the keyboard guideline to an obsolete value if it was moved by
the built-in IME animation in between.